solve issues due to running ccminer on a machine having a cpu with more than 32threads

+++ ccminer.cpp
@@ -119,7 +119,7 @@ static const bool opt_time = true;
 volatile enum sha_algos opt_algo = ALGO_AUTO;
 int opt_n_threads = 0;
 int gpu_threads = 1;
-int64_t opt_affinity = -1L;
+uint64_t opt_affinity = -1L;
 int opt_priority = 0;
 static double opt_difficulty = 1.;
 bool opt_extranonce = true;
@@ -540,12 +540,12 @@ static inline void drop_policy(void) {
 #endif
 }
 
-static void affine_to_cpu_mask(int id, unsigned long mask) {
+static void affine_to_cpu_mask(int id, uint64_t mask) {
 	cpu_set_t set;
 	CPU_ZERO(&set);
 	for (uint8_t i = 0; i < num_cpus; i++) {
 		// cpu mask
-		if (mask & (1UL<<i)) { CPU_SET(i, &set); }
+		if (mask & (1ULL<<i)) { CPU_SET(i, &set); }
 	}
 	if (id == -1) {
 		// process affinity
@@ -558,17 +558,17 @@ static void affine_to_cpu_mask(int id, unsigned long mask) {
 #elif defined(__FreeBSD__) /* FreeBSD specific policy and affinity management */
 #include <sys/cpuset.h>
 static inline void drop_policy(void) { }
-static void affine_to_cpu_mask(int id, unsigned long mask) {
+static void affine_to_cpu_mask(int id, uint64_t mask) {
 	cpuset_t set;
 	CPU_ZERO(&set);
 	for (uint8_t i = 0; i < num_cpus; i++) {
-		if (mask & (1UL<<i)) CPU_SET(i, &set);
+		if (mask & (1ULL<<i)) CPU_SET(i, &set);
 	}
 	cpuset_setaffinity(CPU_LEVEL_WHICH, CPU_WHICH_TID, -1, sizeof(cpuset_t), &set);
 }
 #elif defined(WIN32) /* Windows */
 static inline void drop_policy(void) { }
-static void affine_to_cpu_mask(int id, unsigned long mask) {
+static void affine_to_cpu_mask(int id, uint64_t mask) {
 	if (id == -1)
 		SetProcessAffinityMask(GetCurrentProcess(), mask);
 	else
@@ -576,7 +576,7 @@ static void affine_to_cpu_mask(int id, unsigned long mask) {
 }
 #else /* Martians */
 static inline void drop_policy(void) { }
-static void affine_to_cpu_mask(int id, uint8_t mask) { }
+static void affine_to_cpu_mask(int id, uint64_t mask) { }
 #endif
 
 static bool get_blocktemplate(CURL *curl, struct work *work);
@@ -1890,13 +1890,13 @@ static void *miner_thread(void *userdata)
 		if (opt_affinity == -1L && opt_n_threads > 1) {
 			if (opt_debug)
 				applog(LOG_DEBUG, "Binding thread %d to cpu %d (mask %x)", thr_id,
-						thr_id % num_cpus, (1UL << (thr_id % num_cpus)));
-			affine_to_cpu_mask(thr_id, 1 << (thr_id % num_cpus));
+						thr_id % num_cpus, (1ULL << (thr_id % num_cpus)));
+			affine_to_cpu_mask(thr_id, 1ULL << (thr_id % num_cpus));
 		} else if (opt_affinity != -1L) {
 			if (opt_debug)
 				applog(LOG_DEBUG, "Binding thread %d to cpu mask %lx", thr_id,
-						(long) opt_affinity);
-			affine_to_cpu_mask(thr_id, (unsigned long) opt_affinity);
+						(uint64_t) opt_affinity);
+			affine_to_cpu_mask(thr_id, (uint64_t) opt_affinity);
 		}
 	}
 
@@ -3668,7 +3668,7 @@ void parse_arg(int key, char *arg)
 	case 1020:
 		p = strstr(arg, "0x");
 		ul = p ? strtoul(p, NULL, 16) : atol(arg);
-		if (ul > (1UL<<num_cpus)-1)
+		if (ul > (1ULL<<num_cpus)-1)
 			ul = -1L;
 		opt_affinity = ul;
 		break;
